+++ spatial-data/lib/meta.js
@@ -1,0 +1,69 @@
+const { openDatabase } = require("./db.js");
+
+const { metadata: counties } = require("../sources/counties.js");
+const { metadata: cwas } = require("../sources/countyWarningAreas.js");
+const { metadata: places } = require("../sources/places.js");
+const { metadata: states } = require("../sources/states.js");
+
+const targets = {
+  counties,
+  cwas,
+  places,
+  states,
+};
+
+module.exports = async () => {
+  console.log("fetching versioning metadata...");
+  const db = await openDatabase();
+
+  await db.query(
+    `CREATE TABLE IF NOT EXISTS
+      weathergov_geo_metadata
+      (
+        table_name varchar(512) NOT NULL PRIMARY KEY,
+        version SMALLINT UNSIGNED
+      )`,
+  );
+
+  const existing = await db
+    .query("SELECT * FROM weathergov_geo_metadata")
+    .then((list) =>
+      list.reduce(
+        (o, meta) => ({
+          ...o,
+          [meta.table_name]: meta.version,
+        }),
+        {},
+      ),
+    );
+
+  await db.end();
+
+  const results = {};
+  for (const [target, metadata] of Object.entries(targets)) {
+    results[target] = {
+      update:
+        existing[metadata.table] !== metadata?.version &&
+        metadata?.version > existing[metadata.table],
+    };
+  }
+
+  return results;
+};
+
+module.exports.update = async () => {
+  const db = await openDatabase();
+
+  // eslint-disable-next-line no-restricted-syntax
+  for await (const metadata of Object.values(targets)) {
+    console.log(`setting ${metadata.table} to version ${metadata.version}`);
+    const sql = `INSERT INTO weathergov_geo_metadata
+                  (table_name, version)
+                VALUES("${metadata.table}", "${metadata.version}")
+                ON DUPLICATE KEY
+                  UPDATE version="${metadata.version}"`;
+    await db.query(sql);
+  }
+
+  await db.end();
+};
